Move notification preferences to profile
