+++ app/grandchallenge/notifications/signals.py
@@ -1,16 +1,7 @@
 from actstream import action
-from django.contrib.auth import get_user_model
 from django.db.models.signals import post_save
 from django.dispatch import receiver
 from machina.apps.forum_conversation.models import Topic
-
-from grandchallenge.notifications.models import NotificationPreference
-
-
-@receiver(post_save, sender=get_user_model())
-def create_notification_preferences(sender, *, instance, created, **__):
-    if created:
-        NotificationPreference.objects.create(user=instance)
 
 
 @receiver(post_save, sender=Topic)
