+++ app/grandchallenge/profiles/models.py
@@ -1,8 +1,10 @@
 from hashlib import md5
 from urllib.parse import urlencode
 
+from actstream.models import user_stream
 from django.conf import settings
 from django.contrib.auth import get_user_model
+from django.contrib.contenttypes.models import ContentType
 from django.db import models
 from django.db.models.signals import post_save
 from django.utils.translation import gettext_lazy as _
@@ -49,6 +51,18 @@ class UserProfile(models.Model):
         help_text="Display the organizations that you are a member of in your profile.",
     )
 
+    receive_notification_emails = models.BooleanField(
+        default=True,
+        help_text="Whether to receive email updates about notifications",
+    )
+    notification_email_last_sent_at = models.DateTimeField(
+        default=None, null=True, editable=False
+    )
+    has_notifications = models.BooleanField(default=False, editable=False)
+    notifications_last_read_at = models.DateTimeField(
+        default=None, null=True, editable=False
+    )
+
     def save(self, *args, **kwargs):
         adding = self._state.adding
 
@@ -82,6 +96,19 @@ class UserProfile(models.Model):
                 {"d": "identicon", "s": str(settings.PROFILES_MUGSHOT_SIZE)}
             )
             return gravatar_url
+
+    @property
+    def notifications(self):
+        notifications = user_stream(obj=self.user)
+
+        # Workaround for
+        # https://github.com/justquick/django-activity-stream/issues/482
+        notifications = notifications.exclude(
+            actor_content_type=ContentType.objects.get_for_model(self.user),
+            actor_object_id=self.user.pk,
+        )
+
+        return notifications
 
 
 @disable_for_loaddata
