+++ app/tests/profiles_tests/test_models.py
@@ -7,11 +7,11 @@ from tests.factories import UserFactory
 def test_notification_preference_created():
     u = UserFactory()
 
-    prefs = u.notification_preference
+    prefs = u.user_profile
 
     assert prefs
     assert prefs.user == u
     assert prefs.receive_notification_emails is True
-    assert prefs.email_last_sent_at is None
+    assert prefs.notification_email_last_sent_at is None
     assert prefs.has_notifications is False
-    assert prefs.notifications_last_read_at
+    assert prefs.notifications_last_read_at is None
